fileserver-delete

+++ fileServer/util.js
@@ -64,7 +64,6 @@ async function isDirectoryEmpty(dirPath) {
 async function delDir(p) {
     // 读取文件夹中所有文件及文件夹
     const list = await fse.readdir(p);
-
     for (const v of list) {
         const url = p + '/' + v;
         const stats = await fse.stat(url);
@@ -72,7 +71,7 @@ async function delDir(p) {
         // 判断是文件还是文件夹
         if (stats.isFile()) {
             // 当前为文件，则删除文件
-            await fs.unlink(url);
+            await fs.unlink(url,()=>{});
         } else {
             // 当前为文件夹，则递归调用自身
             await delDir(url);
@@ -85,26 +84,35 @@ async function delDir(p) {
 
 
 async function deleteFilesAndFolder(uploadsDir, maxAge) {
+    
     try {
         // 获取文件夹中的所有文件和目录
         const files = await fse.readdir(uploadsDir);
-
         // 遍历文件和目录
         for (const file of files) {
+            console.log("exp:",file);
             const filePath = path.join(uploadsDir, file);
-            const stats = await fse.lstat(filePath);
-
-            // 如果是文件，并且其修改时间超过最大年龄
-            if (stats.isFile() && moment(stats.mtime).isBefore(moment().subtract(maxAge, 'seconds'))) {
-                // 删除文件
-                await fse.unlink(filePath);
-                console.log(`已删除旧文件: ${filePath}`);
+            if (!fs.existsSync(filePath)) {
+                continue;
             }
+            try{
+                const stats = await fse.lstat(filePath);
+            
+                // 如果是文件，并且其修改时间超过最大年龄
+                if (stats.isFile() && moment(stats.mtime).isBefore(moment().subtract(maxAge, 'seconds'))) {
+                    // 删除文件
+                    await fse.unlink(filePath);
+                    console.log(`已删除旧文件: ${filePath}`);
+                }
 
-            // 如果是目录，并且其修改时间超过最大年龄
-            if (stats.isDirectory() && moment(stats.mtime).isBefore(moment().subtract(maxAge, 'seconds'))) {
-                delDir(filePath);
+                // 如果是目录，并且其修改时间超过最大年龄
+                if (stats.isDirectory() && moment(stats.mtime).isBefore(moment().subtract(maxAge, 'seconds'))) {
+                    delDir(filePath);
+                }
+            }catch(err){
+                continue;
             }
+            
         }
     } catch (err) {
         console.error('删除文件和文件夹时发生错误:', err);
@@ -118,9 +126,31 @@ function clearnUp(dictionary, maxAge, timeInterval) {
     const timer = setInterval(ft, timeInterval);
 }
 
+
+async function movetoBin(filepath,relativepath,binpath)
+{
+    // 确保目标目录存在，如果不存在则创建它
+  if (!fs.existsSync(binpath)) {
+    fs.mkdirSync(binpath, { recursive: true });
+  }
+
+  // 构建目标文件的完整路径
+  const destPath = path.join(binpath, relativepath);
+
+  // 确保目标路径的目录存在
+  const destDir = path.dirname(destPath);
+  if (!fs.existsSync(destDir)) {
+    fs.mkdirSync(destDir, { recursive: true });
+  }
+
+  // 移动文件
+  fs.renameSync(filepath, destPath);
+}
+
 module.exports = {
     formatFileSize: formatFileSize,
     getPrelook: getPrelook,
     clearUploadsDirectory: clearnUp,
-    delDir:delDir
+    delDir:delDir,
+    movetoBin:movetoBin
 }
